Delete a large number of unused, deprecated types
