--- internal/pkg/types/helpers.go
@@ -1,1 +1,0 @@
-package types
